refactor: split ProjectLevelOperations to multiple files
